poc/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "opensdlc-poc"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
anthropic = ["anthropic>=0.40.0", "langchain-anthropic>=0.2.0"]
google = ["google-genai>=1.0.0"]
openai = ["openai>=1.0.0"]
all-llm = ["anthropic>=0.40.0", "google-genai>=1.0.0", "openai>=1.0.0"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]
